[#56341206] Change fork to false for debian
The fork setting was set to true when rhel was added as a potential
target for this library. This causes issues when using older versions of
mongo with upstart.

As this fixes the need for the 'expect daemon' in the upstart config I
have set the default for the param to false and altered the related
tests.