Update Module Version
This relates to https://github.com/alphagov/ci-puppet/pull/430
To install MongoDB 3.x on our 'ci-slaves' we'll need to use a more recent version of the Puppetlabs module.
This version '0.13.0' is the most recent version we can use without breaking our dependencies.  The lastest at the time of writing
this commit being '0.14.0'.

The reason for keeping this module forked is that it has a bug "https://tickets.puppetlabs.com/browse/MODULES-2532" which
Puppetlabs do not seem to want to resolve.